requirements.txt
#--extra-index-url https://pypi.org/simple/
#--trusted-host pypi.org

#timm
#albumentations
#patool
#reformer-pytorch
#sktime
#PyWavelets


ml-trainer-sdk==0.2.4.3
python-dotenv
# aipmodel==0.2.35
aipmodel==0.2.40
boto3
huggingface-hub>=0.34.0,<1.0
clearml
triton==3.2.0
# ryn-data==0.1.3
ryn-data==0.1.6
# opencv-python-headless==4.12.0.88
# opencv-python==4.12.0.88